# docs/swagger.yaml
definitions:
  request.AddProductRequest:
    properties:
      price:
        type: number
      product_name:
        type: string
      quantity:
        type: integer
    required:
    - product_name
    type: object
  request.UpdateProductRequest:
    properties:
      id:
        type: integer
      price:
        type: number
      product_name:
        type: string
      quantity:
        type: integer
    required:
    - id
    type: object
  response.DefaultResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  response.HealthCheckResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Endpoints
  title: CRUD
  version: "1.0"
paths:
  /api/v1/product:
    get:
      consumes:
      - application/json
      description: Get all products from database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      summary: Get all products
    post:
      consumes:
      - application/json
      description: Add a product to database
      parameters:
      - description: product you want to add
        in: body
        name: products
        required: true
        schema:
          $ref: '#/definitions/request.AddProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      summary: Add a product
  /api/v1/product/{productId}:
    delete:
      description: Delete product by its ID
      parameters:
      - description: product ID
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      summary: Delete product by its ID
    get:
      description: Get product by its ID
      parameters:
      - description: product ID
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      summary: Get product by its ID
    patch:
      description: Update product by its ID
      parameters:
      - description: product ID
        in: path
        name: productId
        required: true
        type: string
      - description: things you want to update
        in: body
        name: products
        required: true
        schema:
          $ref: '#/definitions/request.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      summary: Update product by its ID
  /healthz:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.HealthCheckResponse'
      summary: Health Check
swagger: "2.0"